+++ components/board_utils/CMakeLists.txt
@@ -1,0 +1,6 @@
+idf_component_register(
+    SRCS "board_utils.c"
+    INCLUDE_DIRS "."
+    REQUIRES esp_driver_i2c
+    PRIV_REQUIRES driver
+)
