+++ components/boards/CMakeLists.txt
@@ -1,6 +1,6 @@
 # During component requirements scanning, CONFIG_ vars are not yet available.
 # We must declare all possible dependencies upfront.
-set(PRIV_REQUIRES driver nvs_flash esp_driver_i2c esp_driver_spi dac dac_tas57xx dac_tas58xx esp_eth bt main)
+set(PRIV_REQUIRES driver nvs_flash esp_driver_i2c esp_driver_spi dac dac_tas57xx dac_tas58xx esp_eth bt main board_utils)
 
 if(CONFIG_BOARD_TARGET_PATH)
     set(boards_dir ${CONFIG_BOARD_TARGET_PATH})
@@ -12,6 +12,6 @@ endif()
 idf_component_register(
     SRC_DIRS "${boards_dir}" "."
     INCLUDE_DIRS "${boards_dir}" "."
-    REQUIRES esp_driver_i2c
+    REQUIRES esp_driver_i2c board_utils
     PRIV_REQUIRES ${PRIV_REQUIRES}
 )
