+++ components/dac_tas57xx/CMakeLists.txt
@@ -8,5 +8,5 @@ idf_component_register(
     SRCS ${srcs}
     INCLUDE_DIRS "."
     REQUIRES dac
-    PRIV_REQUIRES driver esp_driver_i2c boards
+    PRIV_REQUIRES driver esp_driver_i2c board_utils
 )
